Fix the getCameraPosition() method
The mvMatrix matrix purpose is to transform world coordinates into the
camera's own coordinates. The perspective matrix is then used to
transform those coordinates into the canvas space, it doesn't care about
the camera position :-).

To test this change, you can try feeding the return value of
getCameraPosition to setCameraPosition. Previously, there was a slight
move of the camera, now it stays still (minus the potential floating
point rounding error).