Implemented a built-in hoverpick for mouseover highlighting / tool-tip support.
This is an optional feature, enabled by setting viewer.hoverPickEnabled = true

When enabled, setting/unsetting of the current hover highlight state is automatic.

Uses a dedicated State so that underlying styles are preserved.
Also introduces 'hoverpick' event which should be used in preference to 'mousemove'. hoverpick will never fire when model is moving to help minimise performance impact